Move dialogs to src root
