+++ lib/src/main/java/com/zetcode/dialogs/colourdlg.png
