+++ lib/src/main/java/com/zetcode/dialogs/document-open.png
