+++ lib/src/main/java/com/zetcode/dialogs/notes.png
